requirements.txt
google-api-python-client>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.1
pandas>=1.5
plotly>=5.0
PyYAML>=6.0
